Backend ported to to javascript
