--- server/database/connection.ts
@@ -1,16 +1,0 @@
-export {};
-const mongoose = require("mongoose");
-
-const connectDB = async () => {
-    try {
-    const con = await mongoose.connect(process.env.MONGO_URI, {
-    });
-
-    console.log(`MongoDB connected : ${con.connection.host}`);
-    } catch (err) {
-        console.log(err);
-        process.exit(1);
-    }
-};
-
-module.exports = connectDB;
