+++ server/model/admin.js
@@ -1,6 +1,8 @@
 "use strict";
 Object.defineProperty(exports, "__esModule", { value: true });
+
 const mongoose = require("mongoose");
+
 const schema = new mongoose.Schema({
     Email: {
         type: String,
@@ -9,4 +11,3 @@ const schema = new mongoose.Schema({
 });
 const Admin = mongoose.model("Admin", schema);
 module.exports = Admin;
-//# sourceMappingURL=admin.js.map
