--- server/model/subscription.ts
@@ -1,17 +1,0 @@
-export {};
-const mongoose = require("mongoose");
-
-const schema = new mongoose.Schema({
-    bus: {
-        type: Number,
-        required: true,
-    },
-    subscription: {
-        type: String,
-        required: false,
-    },
-});
-
-const Subscription = mongoose.model("Subscription", schema);
-
-module.exports = Subscription;
